refactor: move types from services to lib/types by domain
- Move StoredRewardsData and TokenBalanceData to lib/types/rewards.ts
- Move PerkEntryStatus to lib/types/perks.ts
- Move UserWalletAddresses to lib/types/wallet.ts
- Consolidate duplicate LeaderboardResponse interfaces in lib/types/leaderboard.ts
- Update imports across 8 service files and 1 API route
- Update lib/types/index.ts to re-export new types
- Maintain encapsulation by keeping internal service types local

Resolves P-126: Refactor types organization
Follows coding principles for domain-specific type organization
