+++ lib/types/perks.ts
@@ -1,0 +1,5 @@
+/**
+ * Perk-related types
+ */
+
+export type PerkEntryStatus = "ineligible" | "open" | "entered" | "closed";
